python_service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
pillow==10.3.0
pdf2image==1.17.0
opencv-python-headless==4.9.0.80
numpy==1.26.4
pytesseract==0.3.10
paddleocr==2.7.0.3
paddlepaddle==2.6.1
python-multipart==0.0.9
